7.30 19:05

+++ utils/ffmpeg.py
@@ -281,14 +281,19 @@ def ffmpeg_audio_tracks_process(input_file, output_file, track_list: list, shell
     return ret
 
 
-def gen_ac3_audio(input_file, output_file, all_ac3_list, audio_index_list, shell, add_dolby=True):
+def gen_ac3_audio(input_file, output_file, all_ac3_list, audio_index_list, shell, add_dolby=True, **kwargs):
+    multi_lang = kwargs.get('multi_lang', False)
+    need_get_ac3_list = kwargs.get('need_get_ac3_list', [])
     ret = True
-    if add_dolby and (all_ac3_list or audio_index_list):
+    if add_dolby and (all_ac3_list or audio_index_list or multi_lang):
         transcode_options = ["-vn", "-acodec", "ac3", "-ab", "320k", "-ac", "6"]
-        if all_ac3_list:
-            ac3_list = list(all_ac3_list)
+        if multi_lang:
+            ac3_list = need_get_ac3_list
         else:
-            ac3_list = list(audio_index_list)
+            if all_ac3_list:
+                ac3_list = list(all_ac3_list)
+            else:
+                ac3_list = list(audio_index_list)
 
         '''
          we add all ac3 channels for the 
@@ -380,7 +385,8 @@ def process_video(input_file, output_file, media_info, offset, duration, target_
     is_hevc = media_info['format']['have_hevc']
     have_264_hevc = media_info['format']['have_264_hevc']
     # have_4k = media_info['format']['have_4k']
-
+    if copy_mode:
+        decode_mode = False
     if file_dur != UNKNOW_SIZE and file_dur < offset + duration:
         raise Exception('invalid param file_dur=%f, offset=%f,duration=%f' % (file_dur, offset, duration))
 
@@ -520,7 +526,7 @@ def process_video(input_file, output_file, media_info, offset, duration, target_
         subprocess.check_output(params, shell=shell, stdin=None, stderr=subprocess.STDOUT, timeout=timeout)
     except BaseException as e:
         ret = False
-        logger.warning('transcode failed %s params=[%s]', e, params)
+        logger.warning('transcode video failed %s params=[%s]', e, params)
     if decode_mode and not ret:
         logger.info(f'try again, dont use decode mode')
         ret = process_video(input_file, output_file, media_info, offset, duration, target_lang, shell,
@@ -583,51 +589,126 @@ def process_audio(input_file, output, media_info, shell, timeout=3600, add_dolby
         map(lambda stream: stream['index'] if stream['codec_name'].find('aac') >= 0 else -1, media_info['stream']))
     audio_index_list = set(
         map(lambda stream: stream['index'] if stream['codec_type'].startswith('audio') else -1, media_info['stream']))
+    language_list = set(
+        map(lambda stream: stream['language'] if stream['codec_type'].startswith('audio') else -1,
+            media_info['stream']))
     ac3_index_list.discard(-1)
     eac3_index_list.discard(-1)
     aac_index_list.discard(-1)
     audio_index_list.discard(-1)
+    language_list.discard(-1)
     all_ac3_list = list(ac3_index_list) + list(eac3_index_list)
-
     # 如果aac_index_list为空, 并且all_ac3_list的长度等于所有音轨的数量, 说明只有ac3/eac3音轨
     only_have_ac3 = False
     if not aac_index_list and len(all_ac3_list) == len(audio_index_list):
         only_have_ac3 = True
 
+    multi_lang = True if len(list(language_list)) > 1 else False
+    language_state = {}
+    trans_aac_list, need_get_ac3_list = [], []
+    if multi_lang:
+        for stream in media_info['stream']:
+            if stream['codec_type'].startswith('audio'):
+                if stream['language'] not in language_state.keys():
+                    language_state[stream['language']] = {}
+                if 'has_aac' not in language_state[stream['language']].keys() or \
+                        not language_state[stream['language']]['has_aac'][0]:
+                    language_state[stream['language']]['has_aac'] = (True, stream['index']) if stream[
+                                                                                                   'codec_name'].find(
+                        'aac') >= 0 else (False, 0)
+
+                if 'has_ac3' not in language_state[stream['language']].keys() or \
+                        not language_state[stream['language']]['has_ac3'][0]:
+                    has_ac3 = stream['codec_name'].find(
+                        'ac3') >= 0 and not stream['codec_name'].find('eac3') >= 0
+                    language_state[stream['language']]['has_ac3'] = (
+                        True, stream['index']) if has_ac3 else (False, 0)
+
+                if 'has_eac3' not in language_state[stream['language']].keys() or \
+                        not language_state[stream['language']]['has_eac3'][0]:
+                    language_state[stream['language']]['has_eac3'] = (True, stream['index']) if stream[
+                                                                                                    'codec_name'].find(
+                        'eac3') >= 0 else (False, 0)
+
+                if 'has_other' not in language_state[stream['language']].keys() or \
+                        not language_state[stream['language']]['has_other'][0]:
+                    has_other = stream['codec_name'].find(
+                        'ac3') < 0 and stream['codec_name'].find('aac') < 0
+                    language_state[stream['language']]['has_other'] = (
+                        True, stream['index']) if has_other else (False, 0)
+
+        for key, value in language_state.items():
+            logger.info(f'{key}, {value}')
+            has_aac_tuple = value['has_aac']
+            has_ac3_tuple = value['has_ac3']
+            has_eac3_tuple = value['has_eac3']
+            has_other_tuple = value['has_other']
+            if has_aac_tuple[0] and has_ac3_tuple[0]:
+                both_have = True
+            elif has_aac_tuple[0]:
+                if not has_eac3_tuple[0] and not has_other_tuple[0]:
+                    # 只有aac, 通过aac生成这个语言的ac3
+                    need_get_ac3_list.append(has_aac_tuple[1])
+                elif has_eac3_tuple[0] and not has_other_tuple[0]:
+                    # 除了aac, 只有eac3, 将eac3转换成aac, 并生成ac3
+                    # trans_aac_list.append(has_eac3_tuple[1])
+                    need_get_ac3_list.append(has_eac3_tuple[1])
+                elif has_other_tuple[0] and not has_eac3_tuple[0]:
+                    # 除了aac, 只有其他类型, 转换成aac, 并生成ac3
+                    # trans_aac_list.append(has_other_tuple[1])
+                    need_get_ac3_list.append(has_other_tuple[1])
+            elif has_ac3_tuple[0]:
+                if not has_eac3_tuple[0] and not has_other_tuple[0]:
+                    # 只有ac3, 需要转换成aac, 并生成ac3
+                    trans_aac_list.append(has_ac3_tuple[1])
+                    need_get_ac3_list.append(has_ac3_tuple[1])
+                elif has_eac3_tuple[0] and not has_other_tuple[0]:
+                    # 除了ac3, 只有eac3, 需要转换成aac, 并生成ac3
+                    trans_aac_list.append(has_eac3_tuple[1])
+                # elif has_other_tuple[0] and not has_eac3_tuple[0]:
+                #     # 除了ac3, 只有其他类型, 需要转换成aac
+                #     trans_aac_list.append(has_other_tuple[1])
+            else:
+                # eac3和其他类型会被转换成aac, 需要生成ac3
+                if not has_eac3_tuple[0]:
+                    need_get_ac3_list.append(has_other_tuple[1])
+                else:
+                    need_get_ac3_list.append(has_eac3_tuple[1])
+
     '''
-    1. acodec_mode控制是否转aac, 1代表需要转成aac, 0则不转
-    2. 当只有ac3音轨或者有其他音轨的时候acodec_mode为 1, 需要转成aac.
+    1. 当只有ac3音轨或者有其他音轨的时候转换成aac
+    2. 存在多语言的时候, 每一种语言都需要有aac和ac3音轨
     '''
     audio_state_list = []
     for stream in media_info['stream']:
         if stream['codec_type'].startswith('audio'):
             audio_state_dic = {'audio_index': stream['index'], 'codec_name': stream['codec_name']}
             if only_have_ac3:
-                audio_state_dic['acodec_mode'] = 1
+                audio_state_dic['target_codec_name'] = 'aac'
             else:
                 # 如果有其他音轨或者有eac3音轨存在就需要转成aac
-                if (stream['codec_name'].find('aac') < 0 and stream['codec_name'].find('ac3') < 0) or stream[
-                    'codec_name'].startswith('eac3'):
-                    audio_state_dic['acodec_mode'] = 1
+                has_other_aduio = stream['codec_name'].find('aac') < 0 and stream['codec_name'].find('ac3') < 0
+                if has_other_aduio or stream['codec_name'].startswith('eac3') or stream['index'] in trans_aac_list:
+                    audio_state_dic['target_codec_name'] = 'aac'
                 else:
-                    audio_state_dic['acodec_mode'] = 0
+                    audio_state_dic['target_codec_name'] = stream['codec_name']
             audio_state_dic['origin_bit_rate'] = stream['bit_rate']
             try:
-                audio_state_dic['channels'] = int(stream['channels']) if stream[
-                                                                             'channels'] and 'channels' in stream.keys() else 0
+                audio_state_dic['channels'] = int(stream['channels']) if stream['channels'] and 'channels' in stream.keys() else 0
             except BaseException as e:
                 audio_state_dic['channels'] = 0
-            audio_state_dic['more_six_channel'] = 1 if int(stream['channels']) > SUPPORT_CHANNEL_NUMBER else 0
+            audio_state_dic['language'] = stream['language']
+            # audio_state_dic['more_six_channel'] = 1 if int(stream['channels']) > SUPPORT_CHANNEL_NUMBER else 0
             audio_state_list.append(audio_state_dic)
 
     audio_src_list = []
     audio_process_result = False
     for audio_state in audio_state_list:
+        logger.info(audio_state)
         audio_src_dic = {}
         bit_rate = AAC_BIT_RATE_MINIMUM
-        audio_src_dic['target_codec_name'] = 'aac' if audio_state['acodec_mode'] == 1 else audio_state['codec_name']
         audio_tracks_list = [audio_state['audio_index']]
-        audio_tmp = f'{output}_{audio_src_dic["target_codec_name"]}_{audio_state["audio_index"]}.tmp'
+        audio_tmp = f'{output}_{audio_state["target_codec_name"]}_{audio_state["audio_index"]}.tmp'
         # bit_mode = audio_state['bit_mode']
         channels = int(audio_state['channels'])
         origin_bit_rate = int(audio_state['origin_bit_rate'])
@@ -636,7 +717,7 @@ def process_audio(input_file, output, media_info, shell, timeout=3600, add_dolby
         aac: 声道为立体声, 原码率小于128修改为128, 大于128小于400修改为192, 大于400修改为320
         ac3: 两声道的码率为192, 大于等于六声道修改为六声道码率为320, 声道数不知道则为六声道码率为320
         '''
-        if audio_src_dic["target_codec_name"].find('aac') >= 0:
+        if audio_state["target_codec_name"].find('aac') >= 0:
             channels = 2
             if origin_bit_rate < AAC_BIT_RATE_MINIMUM:
                 bit_rate = BITE_RATE_MIINIMUM
@@ -644,28 +725,35 @@ def process_audio(input_file, output, media_info, shell, timeout=3600, add_dolby
                 bit_rate = BITE_RATE_MEDIUM
             else:
                 bit_rate = BITE_RATE_MAX
-        if audio_src_dic["target_codec_name"].find('ac3') >= 0:
+        if audio_state["target_codec_name"].find('ac3') >= 0:
             if not channels:
-                channels = 6
+                channels = SUPPORT_CHANNEL_NUMBER
                 bit_rate = BITE_RATE_MAX
             else:
                 if channels <= 2:
                     channels = 2
                     bit_rate = BITE_RATE_MEDIUM
-                elif channels >= 6:
-                    channels = 6
+                elif channels >= SUPPORT_CHANNEL_NUMBER:
+                    channels = SUPPORT_CHANNEL_NUMBER
                     bit_rate = BITE_RATE_MAX
 
         audio_process_result = ffmpeg_audio_tracks_process(input_file, audio_tmp, audio_tracks_list, shell,
                                                            bit_rate=bit_rate, timeout=timeout, channels=channels,
-                                                           target_codec=audio_src_dic["target_codec_name"])
+                                                           target_codec=audio_state["target_codec_name"])
         if audio_process_result:
             # 检查转码完成的文件时长是否异常
             check_audio_ret = check_file_duration(audio_tmp, video_duration)
+            # 多语言的情况下, 可能出现其中有一个语言的时长是正确的, 需要保留正确的那个
+            if multi_lang:
+                # 检测时长和视频不一致, 并且这个音轨需要生成ac3的话, 需要从生成ac3的列表中删除此音轨
+                if not check_audio_ret and audio_state['audio_index'] in need_get_ac3_list:
+                    need_get_ac3_list.pop(need_get_ac3_list.index(audio_state['audio_index']))
+                    continue
             if not check_audio_ret:
-                audio_process_result = False
-                break
+                    audio_process_result = False
+                    break
             audio_src_dic['src'] = audio_tmp
+            audio_src_dic['target_codec_name'] = audio_state["target_codec_name"]
             audio_src_list.append(audio_src_dic)
         else:
             break
@@ -676,13 +764,12 @@ def process_audio(input_file, output, media_info, shell, timeout=3600, add_dolby
                 if audio_src_list[j]['target_codec_name'] != 'aac':
                     audio_src_list[j], audio_src_list[j + 1] = audio_src_list[j + 1], audio_src_list[j]
 
-        # 当只有ac3或者没有ac3音轨的时候, 添加ac3 channel
         gen_ac3_ret = False
-        if not all_ac3_list or only_have_ac3:
+        if not all_ac3_list or only_have_ac3 or multi_lang:
             gen_ac3_ret = gen_ac3_audio(input_file, output_ac3_tmp, all_ac3_list, audio_index_list, shell,
-                                        add_dolby=add_dolby)
+                                        add_dolby=add_dolby, multi_lang=multi_lang, need_get_ac3_list=need_get_ac3_list)
         if gen_ac3_ret:
-            audio_src_list.append({'target_codec_name': 'ac3', 'src': output_ac3_tmp})
+            audio_src_list.append({'src': output_ac3_tmp, 'target_codec_name': 'ac3'})
 
     return audio_src_list, audio_process_result
 
@@ -714,6 +801,7 @@ def transcode(src, output, gpu=True, offset=0.0, duration=0.0, copy_mode=False, 
     :param add_dolby: whether to add dolby
     :return:
     '''
+    failed_reason = ''
     output_full_name = output
     output_tmp = output + '.tmp'
     video_tmp = output + '.video'
@@ -733,6 +821,7 @@ def transcode(src, output, gpu=True, offset=0.0, duration=0.0, copy_mode=False, 
     audio_src_list = []
     audio_process_result = False
     if video_ret:
+        video_duration = int(float(ffprobe_get_media_duration(video_tmp)))
         audio_src_list, audio_process_result = process_audio(src, output, media_info, shell, timeout=timeout,
                                                              add_dolby=add_dolby, video_duration=video_duration)
     if audio_src_list and isinstance(audio_src_list[0], dict):
@@ -743,19 +832,26 @@ def transcode(src, output, gpu=True, offset=0.0, duration=0.0, copy_mode=False, 
         combine_ret = ffmpeg_combine_video_audio(video_src, audio_src_list, output_tmp, shell)
     else:
         logger.warning(f'process video or audio error, video_ret={video_ret}, audio_ret={audio_process_result}')
+        if video_ret and audio_process_result:
+            failed_reason = 'CombineError'
+        elif video_ret:
+            failed_reason = 'AudioError'
+        else:
+            failed_reason = 'VideoError'
 
     if combine_ret:
         shutil.move(output_tmp, output_full_name)
-    try:
+    if os.path.exists(video_tmp):
         os.remove(video_tmp)
-    except BaseException as e:
-        logger.warning(f'remove video tmp error {e}')
+    else:
+        logger.warning(f'not exists file -> {video_tmp}')
+
     for audio_tmp in audio_src_list:
-        try:
+        if os.path.exists(audio_tmp):
             os.remove(audio_tmp)
-        except BaseException as e:
-            logger.warning(f'remove audio tmp error {e}')
-    return combine_ret
+        else:
+            logger.warning(f'not exists file -> {audio_tmp}')
+    return combine_ret, video_duration, failed_reason
 
 
 def ffmpeg_download(url, output, ffmpeg_path, ts=True, timeout=3600,
